Frozen

+++ Veloly-Backend/Controllers/BikeController.cs
@@ -30,7 +30,9 @@ namespace Veloly_Backend.Controllers
                 Price = (decimal)price,
                 PhotoUrl = photoUrl,
                 LockId = lockId,
-                FreeTime = new List<Tuple<DateTime, DateTime>> { new Tuple<DateTime, DateTime>((DateTime)startTime, (DateTime)endTime) },
+                StartTime = (DateTime)startTime,
+                EndTime = (DateTime)endTime,
+                FreeTime = new List<Tuple<DateTime, int>>(),
             };
             db.Bikes.Add(model);
             db.SaveChanges();
@@ -50,22 +52,24 @@ namespace Veloly_Backend.Controllers
         {
             //var tmodel = db.Bikes.FirstOrDefault(t => t.Id == bikeId);
             var model = db.Bikes.FirstOrDefault(x => x.Id == bikeId);
+            if(model == null) return RedirectToAction("Index", "Home");
             model.UserId = userId == null ? model.UserId : userId;
             model.PhotoUrl = photoUrl == null ? model.PhotoUrl : photoUrl;
             model.Price = price == null ? model.Price : (decimal)price;
             model.LockId = lockId == null ? model.LockId : lockId;
             model.Description = description == null ? model.Description : description;
-            model.FreeTime[0] = new Tuple<DateTime, DateTime>((DateTime)startTime, model.FreeTime.ElementAt(0).Item2);
-            model.FreeTime[model.FreeTime.Count() - 1] = new Tuple<DateTime, DateTime>(model.FreeTime.ElementAt(model.FreeTime.Count() - 1).Item1, (DateTime)endTime);
+            model.StartTime = (DateTime)startTime;
+            model.EndTime = (DateTime)endTime;
             db.SaveChanges();
             var json = new Json { JsonString = new System.Web.Script.Serialization.JavaScriptSerializer().Serialize(model) };
             return View("Json", json);
 
         }
 
-        public ActionResult Delete(int id)
+        public ActionResult Delete(int? id)
         {
             var bike = db.Bikes.FirstOrDefault(x => x.Id == id);
+            if (bike == null) return RedirectToAction("Index", "Home");
             db.Bikes.Remove(bike);
             db.SaveChanges();
             return View("Json", new Json());
